requirements.txt
# filepath: /Users/dhairyagundechia/Downloads/finflow/requirements.txt
# Core dependencies
google-adk==0.5.0  # Google Agent Development Kit
google-cloud-aiplatform>=1.28.0
google-cloud-documentai>=2.16.0
google-cloud-bigquery>=3.10.0
google-cloud-storage>=2.10.0

# Utility libraries
pyyaml>=6.0
python-dotenv>=1.0.0

# Development tools
pytest>=7.3.1
pytest-cov>=4.1.0
black>=23.3.0
mypy>=1.3.0
pylint>=2.17.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.2.0
flake8>=6.0.0
isort>=5.12.0

# Application dependencies
langchain>=0.0.235
langchain-google-vertexai>=0.0.1
pandas>=2.0.1
numpy>=1.24.3
tqdm>=4.65.0
requests>=2.31.0
